backend/requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
python-multipart
requests
beautifulsoup4
pdfplumber
langchain
langchain_openai
langchain_groq
python-dotenv